requirements.txt
torch
transformers
pillow
pandas
ipyleaflet
geopandas
shapely
folium
google-streetview
